// src/core/crud/execution-step-type.ts
export enum ExecutionStepType {
  CREATE_MODULE = 'CREATE_MODULE',
  CREATE_CONTROLLER = 'CREATE_CONTROLLER',
  CREATE_SERVICE = 'CREATE_SERVICE',
  CREATE_REPOSITORY = 'CREATE_REPOSITORY',
  CREATE_ENTITY = 'CREATE_ENTITY',
  CREATE_CREATE_DTO = 'CREATE_CREATE_DTO',
  CREATE_UPDATE_DTO = 'CREATE_UPDATE_DTO',
  ADD_PROVIDER = 'ADD_PROVIDER',
  ADD_CONTROLLER = 'ADD_CONTROLLER',
  ORGANIZE_IMPORTS = 'ORGANIZE_IMPORTS',
}